refactor(anima_train): 抽 models + sampling 到 training/（sister script 公开 API）
ADR 0003 PR-A 第 2 步：

- training/models.py    ← load_anima_model / load_vae / load_text_encoders /
                          ensure_models_namespace（被 anima_daemon / anima_generate /
                          anima_reg_ai 通过 _T.X 直接调用，PR-A re-export 保证 0 改动）
- training/sampling.py  ← sigma 调度（_time_snr_shift / _flow_sigmas_simple /
                          _default_noise_sampler / _sample_er_sde_const_x0）+
                          sample_image（训练/生成共用入口）

anima_train.py 行数 2163 → 1731；migration test 全绿。
